xen: block the domain with a timeout, disable interrupts after
The call to SCHEDOP_block is careful to return immediately if
any events are pending. It also re-enables interrupts before blocking.
When SCHEDOP_block returns, we disable interrupts because we
already know there is work to do.

Configure the domain's timer interrupt before we block.

Finally, rename "block_secs" to "block_nsecs" to show that the
units are _nanoseconds_ (thanks dimosthenis.pediaditakis@cl.cam.ac.uk)

Signed-off-by: David Scott <dave.scott@eu.citrix.com>