xen: link against io-page-xen again
Unfortunately unikernels which don't use any device drivers (no network,
no block) fail to link because:

- mirage-xen depends on xen-gnt
- xen-gnt depends on io-page (to represent page mappings)
- io-page needs either io-page-unix or io-page-xen (or mirage-solo5) to
  link

Unfortunately we can't add io-page-xen as a dependency of xen-gnt because
that code has to work on Unix as well as Xen.

Unfortunately xen-gnt is a dependency of mirage-xen because of the need
to inform the grant table code when the domain is suspended and resumed.

If we can break the link between mirage-xen and xen-gnt then we can remove
io-page-xen again. This will probably involve adding some kind of callback
`pre_suspend` `post_resume` mechanism which `xen-gnt` can register with,
if it is linked.

As a temporary measure, this patch puts the io-page-xen back into mirage-xen.

Signed-off-by: David Scott <dave@recoil.org>